Add exponential retry
